+++ src/udgp/instances/nanoparticles.py
@@ -12,7 +12,7 @@ def lj_cluster_points(n: int):
 
     Referência: https://www-wales.ch.cam.ac.uk/~jon/structures/LJ/tables.150.html
     """
-    points = np.loadtxt(DATA_PATH.joinpath(f"lj_cluster/lj_{n}.txt"))
+    points = np.loadtxt(DATA_PATH.joinpath(f"lj_cluster/lj_{n}.txt"), dtype=np.float64)
 
     return points
 
@@ -21,6 +21,6 @@ def c60():
     """
     Retorna: matriz de coordenadas do C60.
     """
-    points = np.loadtxt(DATA_PATH.joinpath(f"c60/c60.txt"))
+    points = np.loadtxt(DATA_PATH.joinpath(f"c60/c60.txt"), dtype=np.float64)
 
     return points
